+++ objective-c/lessons-from_ios-sdk-development/PRPFirstProjectTweeter/PRPFirstProjectTweeter/PRPViewController.h
@@ -1,0 +1,17 @@
+//
+//  PRPViewController.h
+//  PRPFirstProjectTweeter
+//
+//  Created by Juan Gallardo on 4/6/13.
+//  Copyright (c) 2013 Juan Is Awesome. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface PRPViewController : UIViewController {
+    IBOutlet UIWebView *twitterWebView;
+}
+
+@property (nonatomic, strong) IBOutlet UIWebView *twitterWebView;
+
+@end