+++ objective-c/lessons-from_ios-sdk-development/ViewsAndControllers/Recipes_01/Recipes/PRPRecipe.h
@@ -1,0 +1,25 @@
+/***
+ * Excerpted from "iOS SDK Development",
+ * published by The Pragmatic Bookshelf.
+ * Copyrights apply to this code. It may not be used to create training material, 
+ * courses, books, articles, and the like. Contact us if you are in doubt.
+ * We make no guarantees that this code is fit for any purpose. 
+ * Visit http://www.pragmaticprogrammer.com/titles/adios for more book information.
+***/
+//
+//  PRPRecipe.h
+//  Recipes
+//
+//  Created by Bill Dudney on 6/29/11.
+//  Copyright 2011 Gala Factory Software LLC. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//
+@interface PRPRecipe : NSObject
+
+@property(nonatomic, copy) NSString *title;
+
+@end
+//