+++ objective-c/lessons-from_ios-sdk-development/ViewsAndControllers/Recipes_02/Recipes/PRPAppDelegate.h
@@ -1,0 +1,27 @@
+/***
+ * Excerpted from "iOS SDK Development",
+ * published by The Pragmatic Bookshelf.
+ * Copyrights apply to this code. It may not be used to create training material, 
+ * courses, books, articles, and the like. Contact us if you are in doubt.
+ * We make no guarantees that this code is fit for any purpose. 
+ * Visit http://www.pragmaticprogrammer.com/titles/adios for more book information.
+***/
+//
+//  PRPAppDelegate.h
+//  Recipes
+//
+//  Created by Bill Dudney on 9/14/12.
+//  Copyright (c) 2012 Pragmatic Programmers LLC. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class PRPViewController;
+
+@interface PRPAppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@property (strong, nonatomic) PRPViewController *viewController;
+
+@end