+++ HISTORY.md
@@ -1,6 +1,10 @@
 HISTORY
 ==================
 
+__2013-10-04__
++ Created a folder for the tutorials from the book "iOS SDK Development". the path is objective-c/lessons-from_ios-sdk-development
++ objective-c/lessons-from_ios-sdk-development : Copied /ViewsAndControllers/ into this folder. They are the starting points for chapter 4 of iOS SDK Development.  
+
 __2013-30-03__ Juan Gallardo <j@jgallardo.me> <br />
 Began to code along through the book on jQuery mobile. 
 + JavaScript/lessons-from_jquery-mobile/template.html: Created this very basic template.